requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dominate==2.9.1
Flask==3.1.1
flask-wtf==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.0.1
visitor==0.1.3
Werkzeug==3.1.3
